Update docs, move file
