+++ src/server.ts
@@ -29,7 +29,7 @@ import config from './config'
 import lookupTableRoutes from './lookup-table/routes'
 import snapshotRoutes from './components/snapshots/routes'
 import { routeGetLanguageFile } from './components/localisation/routes'
-import { routeTriggers } from './components/routeTriggers'
+import { routeTriggers } from './components/other/routeTriggers'
 import { extractJWTfromHeader, getTokenData } from './components/permissions/loginHelpers'
 require('dotenv').config()
